backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv

#to install:
#pip install -r requirements.txt
